Booking Project/blog-api/requirements.txt
Django==2.2
django-cors-headers==2.5.3
djangorestframework==3.9.2
Pillow==10.1.0
pytz==2023.3.post1
sqlparse==0.4.4
